% Generated by roxygen2 (4.0.2): do not edit by hand
\name{POSTraster}
\alias{POSTraster}
\title{Posts a raster to a GeoServer}
\usage{
POSTraster(access.point, workspace, coverage.store, raster)
}
\arguments{
\item{access.point}{GeoServer REST API access point}

\item{workspace}{GeoServer workspace}

\item{coverage.store}{GeoServer coverage store}

\item{raster}{raster}
}
\value{
charecter vector
}
\description{
Posts a raster to a GeoServer given an existing workspace and coverage store.
}
\examples{
\dontrun{
filename <- "/Users//fbrito/Downloads/ARMOR3D_REPv3-1_20121121_20140420-2.nc"
r <- raster(filename)
POSTraster("http://localhost:8080/geoserver/rest", workspace="aname", coverage.store="astore", r)
}
}

