% Generated by roxygen2 (4.0.2): do not edit by hand
\name{POSTRequest}
\alias{POSTRequest}
\title{A function to POST content to an access point}
\usage{
POSTRequest(access.point, content.type, content)
}
\value{
a list with three elements: status, statusMessage and Location
}
\description{
A function to POST content to an access point
}
\examples{
\dontrun{
POSTRequest(access.point=data.api, content=toJSON(list(items=json.list)), content.type="application/json")
}
}
\keyword{utilities}

